Moved PollenSample and associated structs to separate file
